+++ examples/example/react/src/examples/descriptions/01/App.tsx
@@ -8,5 +8,5 @@ const items = [
 ]
 
 export default function App() {
-  return <Descriptions title="用户信息" items={items} column={{ xs: 1, md: 2 }} />
+  return <Descriptions title="用户信息" items={items} />
 }