+++ examples/example/vue3/src/examples/list/01/App.vue
@@ -9,5 +9,5 @@ const items = [
 </script>
 
 <template>
-  <List :data-source="items" size="md" bordered="divided" hoverable />
+  <List :data-source="items" size="md" split />
 </template>